+++ apps/prairielearn/src/pages/instructorGradingJob/instructorGradingJob.ts
@@ -8,8 +8,7 @@ import { HttpStatusError } from '@prairielearn/error';
 import * as sqldb from '@prairielearn/postgres';
 
 import { makeS3ClientConfig } from '../../lib/aws.js';
-import { typedAsyncHandler } from '../../lib/res-locals.js';
-import type { ResLocalsForPage } from '../../lib/res-locals.types.js';
+import { type ResLocalsForPage, typedAsyncHandler } from '../../lib/res-locals.js';
 import { selectAndAuthzVariant } from '../../models/variant.js';
 
 import {
